LPS-71121 Don't store the search container in memory since it causes article versions to be counted incorrectly for 70x backport.
When LPS-71121 was committed to master, it removed this code. However, LPS-71537, which added the code, was not backported to ee-7.0.x until after LPS-71121 was. Therefore, LPS-71121 did not remove this code when it was backported, since the code did not exist at the time. Now that LPS-71537 has been backported, we need to remove this code.